Support new module -T CONFIG to update config
Also:
1: Allow -C without -F, which only do factory config reset with out firmware update.
    -T BMC    -C  means only factory reset BMC
    -T BIOS   -C  means only ResetEFIVar
    -T CONFIG -C  means only ResetBIOS
2: Let --show_all_versions display all firmware versions, including the golden_image_config
3: Refactor arg validate functions for different operations.

+++ src/http_accessor_curl.py
@@ -104,8 +104,10 @@ class HTTP_Accessor(object):
 
         resp_body    = self._read_file(resp_body_file)
         resp_headers = self._read_file(resp_headers_file)
-        os.remove(resp_body_file)
-        os.remove(resp_headers_file)
+        if os.access(resp_body_file, os.F_OK):
+            os.remove(resp_body_file)
+        if os.access(resp_headers_file, os.F_OK):
+            os.remove(resp_headers_file)
 
         request  = CURL_Request(self.url, self.method, command)
         response = CURL_Response(resp_body, resp_headers, request)
